updating softlink logic for using custom checkpoints

+++ src/tasks/generate.py
@@ -37,8 +37,8 @@ def hotswap_resolution(json):
 
 
 def softlink_checkpoint(checkpoint_path):
-    unique_id = str(uuid.uuid4())
     filename = os.path.basename(checkpoint_path)
+    [model_title, ext] = os.path.splitext(filename)
 
     softlink_path = f"/workspace/stable-diffusion-webui/models/Stable-diffusion/{filename}"
 
@@ -49,7 +49,7 @@ def softlink_checkpoint(checkpoint_path):
     subprocess.run(["ln", "-s", checkpoint_path, softlink_path])
     print(f"Softlinked user checkpoint to {softlink_path}")
 
-    return softlink_path, unique_id
+    return softlink_path, model_title
 
 
 def refresh_checkpoints():
@@ -74,18 +74,24 @@ def handle_checkpoint(json_data):
     checkpoint_path = json_data.get(
         "override_settings").get("sd_model_checkpoint")
 
-    softlink_path, model_name = softlink_checkpoint(checkpoint_path)
+    softlink_path, model_title = softlink_checkpoint(checkpoint_path)
     checkpoints = refresh_checkpoints()
     pprint(checkpoints)
 
-    try:
-        [match] = [c for c in checkpoints if c['model_name'] == model_name]
-        json_data['override_settings']['sd_model_checkpoint'] = softlink_path
+    # Find the matching checkpoint
+    match = None
+    for c in checkpoints:
+        if c['title'] == model_title:
+            match = c
+            break
+
+    if match is not None:
+        json_data['override_settings']['sd_model_checkpoint'] = model_title
         return json_data, softlink_path
-    except Exception as err:
-        print("Error: ", err)
+    else:
+        print(f"Checkpoint {model_title} not found")
         os.remove(softlink_path)
-        raise Exception(f"Checkpoint {model_name} not found")
+        raise Exception(f"Checkpoint {model_title} not found")
 
 
 def generate_handler(json_data):
